get.index {multtest}	R Documentation

Function to compute indices for ordering hypotheses in Package 'multtest'

Description

The hypotheses tested in a multiple testing procedure (MTP), can be ordered based on the output of that procedure. This function orders hypotheses based on adjusted p-values, then unadjusted p-values (to break ties in adjusted p-values), and finally test statistics (to break remaining ties).

Usage

get.index(adjp, rawp, stat)

Arguments

adjp
Numeric vector of adjusted p-values.

rawp
Numeric vector of unadjusted ("raw") marginal p-values.

stat
Numeric vector of test statistics.

Value

Numeric vector of indices so that the hypotheses can be ordered accroding to significance (smallest p-values and largest test statistics first). This function is used in the plot method for objects of class MTP to order adjusted p-values for graphical summaries. The summary method for objects of class MTP will return these indices as its second component.

Author(s)

Katherine S. Pollard

See Also

MTP, plot,MTP,ANY-method, summary,MTP-method

Examples

data<-matrix(rnorm(200),nr=20)
mtp<-MTP(X=data,test="t.onesamp")
index<-get.index(adjp=mtp@adjp,rawp=mtp@rawp,stat=mtp@statistic)
mtp@statistic[index]
mtp@estimate[index]
apply(data[index,],1,mean)

[Package multtest version 2.40.0 Index]