mt.reject {multtest}	R Documentation

Identity and number of rejected hypotheses

Description

This function returns the identity and number of rejected hypotheses for several multiple testing procedures and different nominal Type I error rates.

Usage

mt.reject(adjp, alpha)

Arguments

adjp
A matrix of adjusted p-values, with rows corresponding to hypotheses and columns to multiple testing procedures. This matrix could be obtained from the function mt.rawp2adjp .

alpha
A vector of nominal Type I error rates.

Value

A list with components

r
A matrix containing the number of rejected hypotheses for several multiple testing procedures and different nominal Type I error rates. Rows correspond to Type I error rates and columns to multiple testing procedures.

which
A matrix of indicators for the rejection of individual hypotheses by different multiple testing procedures for a nominal Type I error rate alpha[1]. Rows correspond to hypotheses and columns to multiple testing procedures.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine,
Yongchao Ge, yongchao.ge@mssm.edu.

See Also

mt.maxT, mt.minP, mt.rawp2adjp, golub.

Examples

# Gene expression data from Golub et al. (1999)
# To reduce computation time and for illustrative purposes, we condider only
# the first 100 genes and use the default of B=10,000 permutations.
# In general, one would need a much larger number of permutations
# for microarray data.

data(golub)
smallgd<-golub[1:100,] 
classlabel<-golub.cl

# Permutation unadjusted p-values and adjusted p-values for maxT procedure
res<-mt.maxT(smallgd,classlabel)
mt.reject(cbind(res$rawp,res$adjp),seq(0,1,0.1))$r

[Package multtest version 2.40.0 Index]